refactor: rename universe enum/model file to standing, update enum docs for corporation.
