+++ src/model/enums/mod.rs
@@ -13,4 +13,4 @@ pub mod character;
 pub mod corporation;
 pub mod market;
 pub mod notification;
-pub mod universe;
+pub mod standing;
