+++ src/model/mod.rs
@@ -24,4 +24,4 @@ pub mod corporation;
 pub mod enums;
 pub mod market;
 pub mod oauth2;
-pub mod universe;
+pub mod standing;
